Safari-609.1.11.1/LayoutTests/accessibility/ios-simulator/focus-change-notifications-expected.txt

H1

H2

This tests that notifications are fired for focus changes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
